feat(comments): make saperate comment controller and its route changes
